% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Splits a string into a vector containing each element individually}
\usage{
strsplit1(string, split)
}
\arguments{
\item{string}{A character vector with one element}

\item{split}{What to split on.}
}
\value{
A character vector containing the elements of the original string
}
\description{
Splits a string into a vector containing each element individually
}
\examples{
string <- "alpha,bravo,charlie,delta"
strsplit1(string, split = ",")
}
